+++ src/api.v2/helpers/pipeline/__mocks__/gem2s.js
@@ -1,6 +1,6 @@
 const startGem2sPipeline = jest.fn();
 const handleGem2sResponse = jest.fn();
-const { runGem2s } = jest.requireActual('../gem2s.js');
+const runGem2s = jest.fn();
 
 module.exports = {
   startGem2sPipeline,
